Ogone: merge ogone_response.rb into ogone.rb
